+++ src/main/factory/usecase/account/db-authentication-factory.ts
@@ -1,8 +1,8 @@
-import { DbAuthentication } from "../../../data/usecases/authentication/db-authentication"
-import { BcryptAdapter } from "../../../infra/criptography/bcrypt/bcrypt-adapter";
-import { JWTAdapter } from "../../../infra/criptography/jwt/jwt-adapter";
-import { AccountMongoRepository } from "../../../infra/db/mongodb/account/account-mongo-repository";
-import env from "../../config/env";
+import { DbAuthentication } from "../../../../data/usecases/authentication/db-authentication"
+import { BcryptAdapter } from "../../../../infra/criptography/bcrypt/bcrypt-adapter";
+import { JWTAdapter } from "../../../../infra/criptography/jwt/jwt-adapter";
+import { AccountMongoRepository } from "../../../../infra/db/mongodb/account/account-mongo-repository";
+import env from "../../../config/env";
 
 export const makeDbAuthentication = (): DbAuthentication => {
     const salt = 12;
